Refactoring ActiveTransactionsQueue
