+++ arangod/Replication2/StateMachines/Document/DocumentFollowerState.h
@@ -23,6 +23,7 @@
 
 #pragma once
 
+#include "Replication2/StateMachines/Document/ActiveTransactionsQueue.h"
 #include "Replication2/StateMachines/Document/DocumentStateMachine.h"
 #include "Replication2/StateMachines/Document/DocumentStateTransactionHandler.h"
